file_converter_project/requirements.txt
Pillow
pdf2docx
docx2pdf
pdf2image
tabula-py
pandas
fpdf2
python-pptx
comtypes